Create a UITextView subclass to handle maximum height when TextView should scroll
Create a ContainerView for comment form including TextView, post button and a placeholder label for the input area.

Create a container for the comment composer that includes avatar of currently logged in user and InputContainerView

Add a label to show the backer's state to cause only backers can post comment.