re-recorded all snapshots with snapshot-testing framework.
This is a huge commit. But only because it's a lot of files. The idea being it replaces everything in the `_64` folder which had our old tests that were references in the `FBSnapshotTestCase` dependency. These new screenshots are not going to be in one folder. They will be local to the view they represent, similar to the organization of the `Kickstarter-Framework-iOS` target file structure. It's more ordered than having one giant reference folder because if you need to check the snapshots, it's easier find any given views' subfolder in XCode. Opening that views subfolder in XCode will give you access to the screenshots in that folder, but through Finder. The image files themselves are not included in the XCode project, so it doesn't bloat our project either.