[NTV-648] Agnostic Snapshot Testing Between M1 and Intel (#1794)
* trial new snapshots on m1 snapshot testing framework with 0.98 perceptualPrecision

* intel failed comparing against m1 recordings of this test. Trying a subview recording.

* replaced FBSnapshotVerifyView with assertSnapshot everywhere in Kickstarter-Framework-iOSTests

* re-recorded all snapshots with snapshot-testing framework.

This is a huge commit. But only because it's a lot of files. The idea being it replaces everything in the `_64` folder which had our old tests that were references in the `FBSnapshotTestCase` dependency. These new screenshots are not going to be in one folder. They will be local to the view they represent, similar to the organization of the `Kickstarter-Framework-iOS` target file structure. It's more ordered than having one giant reference folder because if you need to check the snapshots, it's easier find any given views' subfolder in XCode. Opening that views subfolder in XCode will give you access to the screenshots in that folder, but through Finder. The image files themselves are not included in the XCode project, so it doesn't bloat our project either.

* updating a missed set of tests

* updated tests not to overwrite each other when using the same assert.

* some tests running on Intel did require approximate comparisons, so added perceptual precision to only those classes.

Tests now recorded on M1 pass 100% on Intel.

* added test for facebook reset password view controller

recorded on M1, should pass on CI and locally on Intel.

* removed _64 directory, where old screenshots were, remove ios-snapshot-test-case as dependency, updated TestCase class, removed project schema variables.

* still using old trait controllers, combos, devices, languages and other features. Mainly because we still cannot enforce a simulator to fail a record if its' not iPhone 8.

It'll just have to be common knowledge that iPhone is our recording simulator. Will update the readme. It's already kind of a guarantee because if we record with another simulator, we'd have failing tests one existing tests when the whole suite is run.

* updated snapshots on FacebookResetPasswordViewController

* Added preferred simulator check, so no tests pass if simulator for scheme is not iPhone 8 iOS 14.5. Same for recording.