[MBL-59] Fix personal data export. (#1836)
* [MBL-59] Fix personal data export.

Update `ExportDataEnvelope` `State` types to align with the types defined by the server. Update related logic to be more readable. Also update some test cases to default to `none` instead of `expired`, now that `none` is the default state for users that have never requested their data before.

* [MBL-59] Handle unknown states

Update the ExportDataEnvelope `state` enum to rely on an underlying `stateString`. This allows a default `unknown` value to be used if the state from the server is missing or doesn't match a value in the enum.

* rename stateString to rawState