Fix crash when dark mode is enabled (#2472)
* Fix crash when dark mode is enabled

The HTML initialization method of NSAttributedString isn't thread safe (and, I suspect, also isn't background safe). With dark mode enabled, the app takes a light/dark screenshot pair when the app moves into the background. This triggers a trait collection change, which triggers a change to bindStyles, which ultimately triggered this attributed string to crash.

* Also handle PLOT text